cool#11226 sw per-view redline on: move the setter to the model
Open e.g. an RTF file with \revisions in it (enable track changes
recording), notice that track changes is not enabled, when the UI would
take the "redline on" setting from the view.

Commit 025a1e4612fecf59f38910fbf52fc63db054ae5f (cool#11226 sw per-view
redline on: add new flag in SwViewOption, 2025-02-26) added the per-view
"is record" setter to SwEditShell, but I missed that this setting can be
also set by an import filter, which works on an SwDoc with no view.

Fix this problem by:

- remove the setter from SwEditShell::SetRedlineFlags(), we'll do this
  in the model instead
- use sw::DocumentRedlineManager::SetRedlineFlags_intern() everywhere,
  so we have a central point where the current view (if there is any)
  can be updated together with the model
- implement the model -> current view synchronization in
  sw::DocumentRedlineManager::SetRedlineFlags_intern()
- extend the the SwView ctor, so the setting of the view is not the
  default, but it's based on the model

Also make it explicit that RedlineFlags::Ignore is not something a user
may set. With this rearrangement, implementing a getter (that takes the
view into account) looks possible, but is not yet done in this commit.

Change-Id: Ifc0c3bbb95539cc1a26582ee7d27653899fbc236
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/182302
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Tested-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>