cool#11885 lok: sc: translate TabPrefix by user
In LOKit different users can have different locales on the same file, so
when creating a new sheet with the default name, we must ensure that the
proper localized name is used. Otherwise the language with which the
module is first loaded is always used.

A test cannot be added since to test this multiple languages have to be
enabled and in Jenkins right now only English is enabled.

Change-Id: I6aa4caacc7d079fbf9b14ddf9e27428d77411adf
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185055
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
(cherry picked from commit 76a02119f39ffdf9dde8f6122707da95490f2253)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/185305
Reviewed-by: Andras Timar <andras.timar@collabora.com>